ddddd

+++ vtm-web/src/org/oscim/gdx/emu/org/oscim/utils/Tessellator.java
@@ -1,7 +1,7 @@
 package org.oscim.utils;
 
 import org.oscim.core.GeometryBuffer;
-import org.oscim.renderer.elements.VertexData;
+import org.oscim.renderer.bucket.VertexData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -107,7 +107,7 @@ public class Tessellator {
 		return numIndices;
 	}
 
-	public static int tessellate(float[] points, int ppos, int plen, short[] index,
+	public static int tessellate(float[] points, int ppos, int plen, int[] index,
 	        int ipos, int rings, int vertexOffset, VertexData outTris) {
 
 		Int32Array io;
